+++ src/app/pipes/virtual-scroll/virtualscroll.pipe.ts
@@ -1,0 +1,12 @@
+import { Pipe, PipeTransform } from '@angular/core';
+
+@Pipe({
+  name: 'virtualscroll',
+  pure: false
+})
+export class VirtualscrollPipe implements PipeTransform {
+
+  transform(value: any, start: number, object_max: number): any {
+    return value.slice(Math.max(0, start - 10), start + object_max);
+  }
+}